+++ anthem/lyrics/modules.py
@@ -18,12 +18,25 @@ def uninstall(ctx, module_list):
         raise AnthemError(u'Cannot uninstall modules. See the logs')
 
 
-def update_translations(ctx, module_list):
+def update_translations(ctx, module_list, overwrite=False):
     """ Update translations from module list"""
-    if not module_list:
+    if not isinstance(module_list, list):
         raise AnthemError(u"You have to provide a list of "
                           "module's name to update the translations")
+    if overwrite:
+        ctx.log_line(
+            u'All previous translations will be dropped for requested addons'
+        )
+    ir_module = ctx.env['ir.module.module']
+    if hasattr(ir_module, 'update_translations'):
+        # Odoo version <= 10.0
+        method_name = 'update_translations'
+    else:
+        # Odoo version >= 11.0
+        method_name = '_update_translations'
 
-    for module in module_list:
-        ctx.env['ir.module.module'].with_context(overwrite=True).search(
-            [('name', '=', module)]).update_translations()
+    domain = [('name', 'in', module_list)]
+    mods = ctx.env['ir.module.module'].search(domain)
+    ctx.log_line('Reloading translations for %s' % str(module_list))
+    mods.with_context(overwrite=overwrite)
+    getattr(mods, method_name)()