+++ HISTORY.rst
@@ -8,7 +8,12 @@ Unreleased
 
 **Features**
 
+- BREAKING: Change default `overwrite` value for
+   ``lyrics.modules.update_translations`` to False
+
 **Bugfixes**
+
+- Make ``lyrics.modules.update_translations`` Odoo >= 11.0 compatible
 
 **Improvements**